Final_project/backend/requirements.txt
flask==3.0.2
pydub==0.25.1
SpeechRecognition==3.10.0
googletrans==4.0.0-rc1
gtts==2.4.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0